Generate INSPIRE IDs

Given pairs of coordinates, generates their INSPIRE grid representation. Given INSPIRE identifiers, can also extract the X and Y coordinates.

An INSPIRE ID contains both the cell size and the ETRS89-LAEA coordinates of the south-west corner of the grid cell in the format {cellsize}N{x_coord}E{y_coord}. Only the first four (or five if res = "100m") digits of the coordinates are used in the identifier.

z22_inspire_generate(coords, res = NULL, legacy = FALSE)

z22_inspire_extract(inspire, as_sf = FALSE)

Arguments

coords

A list, matrix, or dataframe where the X and Y coordinates are either in the columns "x" and "y" or in the first and second column position, respectively. Column names are converted to lowercase.

Can also be a sf/sfc object in which case the coordinates are extracted using st_coordinates.

res

Resolution of the grid. Can be "100m", "250m", "1km", "5km", "10km", or "100km". If NULL, tries to guess the resolution from the provided coordinates.

legacy

If TRUE, generates legacy INSPIRE ID. Defaults to FALSE.

inspire

A vector of INSPIRE IDs. Can be either legacy or non-legacy.

as_sf

Whether to return an object of class sfc or a tibble.

Value

z22_inspire_generate returns a character vector containing the INSPIRE identifiers. z22_inspire_extract returns a dataframe or sfc object containing the points extracted from the INSPIRE identifiers.

Details

To remain fast even for huge grid datasets, the function is just a very simple sprintf wrapper that performs no input checks. To produce valid INSPIRE identifiers, make sure to transform your data to ETRS89-LAEA (e.g. using st_transform(..., 3035)). You should also make sure that the coordinates are the south-west corner of existing INSPIRE grid cells.

Examples

# Generate IDs from a dataframe
coords <- data.frame(x = c(4334150, 4334250), y = c(2684050, 2684050))
z22_inspire_generate(coords) |>
  z22_inspire_extract() |>
  identical(coords)
#> [1] FALSE

# Extract coordinates from legacy ID strings
z22_inspire_extract("100mN34000E44000")
#> # A tibble: 1 × 2
#>         x       y
#>     <dbl>   <dbl>
#> 1 4400050 3400050

# Generate IDs from an sf dataframe
if (requireNamespace("sf", quietly = TRUE)) {
  coords <- sf::st_as_sf(coords, coords = c("x", "y"))
  z22_inspire_generate(coords)
}
#> [1] "CRS3035RES100mN26840E43341" "CRS3035RES100mN26840E43342"