Get Census 2022 grid dataset

Retrieve the values and coordinates of gridded features from the censuses 2011 and 2022.

When we are talking about a

feature, we talk about an indicator aggregated to grid cells, e.g., age or the number of dwellings.

category, we talk about the discrete classifications of features, e.g., ages 10 to 19, 20 to 20, 30 to 39, etc.

Both feature and category have to be provided to uniquely identify a dataset.

z22_data(
  feature,
  categories = NULL,
  year = 2022,
  res = "1km",
  all_cells = FALSE,
  normalize = FALSE,
  rasterize = FALSE,
  as_sf = FALSE,
  update_cache = FALSE
)

Arguments

feature

A grid feature. See z22_features for a list of available features. You can pass both English names and legacy names (i.e., variable names from the 2011 census).

categories

One or multiple feature categories. See z22_categories for a list of available categories. If NULL, retrieves all categories for a given feature. Generally, the more categories are selected, the longer the download.

res

Resolution of the grid dataset. Can be "100m", "1km", or "10km". If year is 2011, "10km" is not available and some features are only available at certain resolutions.

all_cells

If TRUE, joins the retrieved attribute with the complete grid from z22_grid. Otherwise, the attribute grid will contain only those grid cells with one or more recorded units. Defaults to FALSE, because loading the grid and joining with it is computationally expensive.

rasterize

If TRUE and the terra package is installed, converts the attribute coordinates to a SpatRaster.

as_sf

If TRUE and the sf package is installed, converts the attribute coordinates to an sf tibble.

cache

By default, both functions cache attribute files for the remainder of the R session. They are downloaded to a temporary directory and - if the file to download already exists - are recovered from the cache. In other words, when rerunning the same request multiple times, the subsequent calls should be much faster. If FALSE, disables caching for this call and overwrites the currently cached attribute file (if any) with a fresh one. Defaults to TRUE, i.e. always cache.

Value

A tibble, SpatRasterDataset or sf tibble depending on the rasterize and as_sf arguments.

If a tibble is returned each category in categories is stored in a column. If a SpatRasterDataset is returned, each category is a named layer.

Details

Half of the grids cell width is added to each coordinate in the dataset internally. According to the INSPIRE guidelines, coordinates always represent the South-west of the grid cells. Centroids represent the geographic location of grid cells better which is why they are used.

Examples

# Get gridded population
pop <- z22_get_attribute("population", all_cells = TRUE, rasterize = TRUE)
#> Error in z22_get_attribute("population", all_cells = TRUE, rasterize = TRUE): could not find function "z22_get_attribute"
terra::plot(pop)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'pop' not found

# Get data about buildings using district heating
z22_get_attribute("buildings", "HEIZTYP", 1)
#> Error in z22_get_attribute("buildings", "HEIZTYP", 1): could not find function "z22_get_attribute"